feat: show readonly navbar
